Inject Spring EL into GraphQL Directive @fetch(from="...")
For language injection to work at all, had to let the PSI class for
GraphQL string values implement PsiLanguageInjectionHost.
Also, using Spring EL requires a dependency on IDEA plugins
com.intellij.spring and com.intellij.jsp. You need both, because
"spring" only has an optional dependency on "jsp" and activates its
Spring EL support only if "jsp" is available.
Unfortunately, it is not possible for an optional plugin dependency
to require *two* plugins to be available, so for the time being, I
added both dependencies as required (not "optional").

After having resolved these issues, it is easy to add language
injection for a GraphQL directive. GraphQLSchemaSpringELInjector
checks for directive name "fetch" and argument name "from".
It might even be possible to extend JetBrains' generic Language
Injection plugin to be able to inject into GraphQL, but that's
another story.